Extract default auto checkout duration logic to extension function
